# Speedup ToTime