--- a/reddo/benchmark_test.go
+++ b/reddo/benchmark_test.go
@@ -1,0 +1,15 @@
+package reddo
+
+import (
+	"testing"
+	"time"
+)
+
+func BenchmarkToTime(b *testing.B) {
+	now := time.Now()
+
+	for i := 0; i < b.N; i++ {
+		_, _ = ToTime(now)
+
+	}
+}