By comparing directly to TypeTime, instead of indirectly go through a few string comparison.

While at it, also gofmt the source.

```
name      old time/op    new time/op    delta
ToTime-8    59.3ns ± 0%    37.5ns ± 2%  -36.88%  (p=0.000 n=9+9)
```