@btnguyen2k 2022-11-09
Hi @cuonglm, while your implementation is correct, the gofmt change is not necessary. The original source is complying with gofmt -s as it uses one tab while yours uses two spaces - which is not really a fix.
Can you please revert the gofmt change?
Also, please remove the benchmark file or rework on it as it is not well named.

Best regards,
Thanh