@cuonglm 2022-11-09
@btnguyen2k What's `gofmt` version you are using?

```
~/sdk/go1.19.3/bin/gofmt -s -d .
```

reformat the file as of this CL.

> Also, please remove the benchmark file or rework on it as it is not well named.

What's not well named?