@btnguyen2k 2022-11-09
@cuonglm I am using v1.17.x, didn't know v1.19.x behaves differently. So this gofmt fix is accepted.